+++ proj2/algorithms.h
@@ -1,5 +1,5 @@
 #pragma once
 
-int sieve(int max_number);
+int sieve(unsigned long max_number);
 
-int sieve_no_even(int max_number);
+int sieve_no_even(unsigned long max_number);
